Removed multiple inheritance handling (it was not complete anyway) and cleaned up. More straightforward compilation to bytecode (not method for field access).
Better code generated for incrementation (preincrementation should be changed too).

More precise type for integer literals (allows for "short s = 1", that is 1 has type byte and not int). This is not exactly JLS (where 1 has type byte, but special assignment typecheking is done: section 5.2). I'm not sure it makes a difference though.