Complete support for dynamic link mode (nicec -d).
Added info in bytecode to get the correct definition of an alternative
(used to incorrectly assume if was in the same package).

Better handling of late added native classes
(used to create exceptions. are now ignored. the corresponding reflect method is also ignored, which could be surprising to the user).