Massive update of gnu.*
These packages have been resynchonised with the upstream source (Kawa).
They are now under a 'double-repository' policy, so that they can be kept easily up-to-date while keeping the local differences.

The bytecode generation phase has also changed greatly:
we create a (list of) gnu.expr.Package instance(s).
All the information for code generation is thus held under a single object.
This allows for a cleaner front-end with much less knowledge of the back-end.