Massive update of gnu.*
The bytecode generation phase has changed greatly:
we create a (list of) gnu.expr.Package instance(s).
All the information for code generation is thus held under a single object.
This allows for a cleaner front-end with much less knowledge of the back-end.

The code is now always generated in directories.
If anoutput archive is requested (with -o file.jar) an additional pass collects classes and writes them in the archive.
This has a clear semantics, independant of which packages are up-to-date.
This also ensures that up-to-date packages are found and thus not recompiled.