Build non-native debian packages.
The CVS repository is now tagged with version_x_y_z for version x.y.z.

A self-contained .tar.gz source archive is now produced, which allows a clean
bootstrap, without depending on an already installed version of Nice.
That archive constitutes the official release of Nice. For instance, the
debian source package can now be defined as a diff with that archive.