Clarified the terminology: we should always use "to declare and to implement" a method, and "a method declaration, a method implementation". The tutorial used to say "to declare and to define" (which is easier to confuse) and "a method declaration, a method alternative". If these uses stay somewhere, they should be updated.
Changed the order, to first present everything about methods, and then
move to advanced typing (parametric classes, precise types).
Added a link to the new tutorial in wiki about methods.

Removed advanced things about interfacing with Java. This does not belong to
the tutorial, and is covered in the manual (added a link).

Added link to the roadmap, about visibility modifiers.